JavaQuestionServiceTest готов

+++ src/main/java/com/agorohov/skyprocw2examservice/service/JavaQuestionService.java
@@ -1,5 +1,8 @@
 package com.agorohov.skyprocw2examservice.service;
 
+import com.agorohov.skyprocw2examservice.exception.ParamIsNotPresentException;
+import com.agorohov.skyprocw2examservice.exception.QuestionInNotExistException;
+import com.agorohov.skyprocw2examservice.exception.QuestionIsAlreadyExistException;
 import com.agorohov.skyprocw2examservice.model.Question;
 import com.agorohov.skyprocw2examservice.repository.QuestionRepository;
 import org.springframework.stereotype.Service;
@@ -20,17 +23,27 @@ public class JavaQuestionService implements QuestionService {
 
     @Override
     public Question add(String question, String answer) {
+        checkParamIsPresent(question);
+        checkParamIsPresent(answer);
         return add(new Question(question, answer));
     }
 
     @Override
     public Question add(Question question) {
+        checkParamIsPresent(question);
+        if (questionRepository.isQuestionPresent(question)) {
+            throw new QuestionIsAlreadyExistException("Такой вопрос уже добавлен, повторное добавление невозможно");
+        }
         questionRepository.addQuestion(question);
         return question;
     }
 
     @Override
     public Question remove(Question question) {
+        checkParamIsPresent(question);
+        if (!questionRepository.isQuestionPresent(question)) {
+            throw new QuestionInNotExistException("Такого вопроса нет, удаление невозможно");
+        }
         questionRepository.removeQuestion(question);
         return question;
     }
@@ -42,6 +55,13 @@ public class JavaQuestionService implements QuestionService {
 
     @Override
     public Question getRandomQuestion() {
-        return questionRepository.getQuestion(random.nextInt(questionRepository.getSize()));
+        return questionRepository.getQuestion(
+                random.nextInt(questionRepository.getSize()));
+    }
+
+    private void checkParamIsPresent(Object o) {
+        if (o == null || o.toString().isBlank()) {
+            throw new ParamIsNotPresentException("Некорректное значение параметра");
+        }
     }
 }
